move state management to appropriate components
