+++ src/pages/workspaces/workspace/Dashboard/CloudInformation.ts
@@ -1,5 +1,5 @@
 import { cond } from '@terra-ui-packages/core-utils';
-import { Fragment, ReactNode } from 'react';
+import { Fragment, ReactNode, useEffect, useState } from 'react';
 import { div, dl, h } from 'react-hyperscript-helpers';
 import { bucketBrowserUrl } from 'src/auth/auth';
 import { ClipboardButton } from 'src/components/ClipboardButton';
@@ -8,8 +8,10 @@ import { icon } from 'src/components/icons';
 import { TooltipCell } from 'src/components/table';
 import { ReactComponent as GcpLogo } from 'src/images/gcp.svg';
 import { Ajax } from 'src/libs/ajax';
+import { withErrorReporting } from 'src/libs/error';
 import Events, { extractWorkspaceDetails } from 'src/libs/events';
-import { newTabLinkProps } from 'src/libs/utils';
+import { useCancellation } from 'src/libs/react-utils';
+import { formatBytes, newTabLinkProps } from 'src/libs/utils';
 import {
   AzureWorkspace,
   canWrite,
@@ -25,8 +27,6 @@ import { InfoRow } from 'src/pages/workspaces/workspace/Dashboard/InfoRow';
 interface CloudInformationProps {
   storageDetails: StorageDetails;
   workspace: Workspace;
-  storageCost?: { isSuccess: boolean; estimate: string; lastUpdated?: string };
-  bucketSize?: { isSuccess: boolean; usage: string; lastUpdated?: string };
 }
 
 interface AzureCloudInformationProps extends CloudInformationProps {
@@ -74,9 +74,52 @@ const AzureCloudInformation = (props: AzureCloudInformationProps): ReactNode => 
 };
 
 const GoogleCloudInformation = (props: GoogleCloudInformationProps): ReactNode => {
-  const { workspace, storageDetails, storageCost, bucketSize } = props;
+  const { workspace, storageDetails } = props;
   const { accessLevel } = workspace;
-  const { googleProject, bucketName } = workspace.workspace;
+  const { bucketName, googleProject } = workspace.workspace;
+
+  const signal = useCancellation();
+
+  const [storageCost, setStorageCost] = useState<{ isSuccess: boolean; estimate: string; lastUpdated?: string }>();
+  const [bucketSize, setBucketSize] = useState<{ isSuccess: boolean; usage: string; lastUpdated?: string }>();
+
+  useEffect(() => {
+    const { namespace, name } = workspace.workspace;
+
+    const loadStorageCost = withErrorReporting('Error loading storage cost data', async () => {
+      try {
+        const { estimate, lastUpdated } = await Ajax(signal)
+          .Workspaces.workspace(namespace, name)
+          .storageCostEstimate();
+        setStorageCost({ isSuccess: true, estimate, lastUpdated });
+      } catch (error) {
+        if (error instanceof Response && error.status === 404) {
+          setStorageCost({ isSuccess: false, estimate: 'Not available' });
+        } else {
+          throw error;
+        }
+      }
+    });
+
+    const loadBucketSize = withErrorReporting('Error loading bucket size.', async () => {
+      try {
+        const { usageInBytes, lastUpdated } = await Ajax(signal).Workspaces.workspace(namespace, name).bucketUsage();
+        setBucketSize({ isSuccess: true, usage: formatBytes(usageInBytes), lastUpdated });
+      } catch (error) {
+        if (error instanceof Response && error.status === 404) {
+          setBucketSize({ isSuccess: false, usage: 'Not available' });
+        } else {
+          throw error;
+        }
+      }
+    });
+
+    if (workspace.workspaceInitialized && canWrite(accessLevel)) {
+      loadStorageCost();
+      loadBucketSize();
+    }
+  }, [workspace, accessLevel, signal]);
+
   return h(Fragment, [
     dl([
       h(InfoRow, { title: 'Cloud Name' }, [
